+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 35.4.0
+New:
+- `GeoDropdownCompactButton`: Disabled state is added to the button.
+- `GeoBorderedToken`: Disabled state is added to the token.
 ## 35.3.1
 Changed:
 - `GeoTree`: A load more button is added. If `hasLoadMoreButton` prop is true, there is a load more button by showing items according to the `pageSize`prop.